requirements.txt
# requirements.txt
openai
python-dotenv
pandas
streamlit
flake8